Tidying and test improvements; fix precision-related bug in applying SOP.
