+++ docs/vulnerability_functions/inundation_jrc/raw.csv
@@ -35,15 +35,15 @@ Transport,3,1,,1,1,,,1,,,,0,,,
 Transport,4,1,,1,1,,,1,,,,0,,,
 Transport,5,1,,1,1,,,1,,,,0,,,
 Transport,6,1,,1,1,,,1,,,,0,,,
-Infrastructure  roads,0,0,,,0,,,0,,,,0,,,
-Infrastructure  roads,0.5,0.25,,,0.214436906,,,0.232218453,,,,0.069410857,,,
-Infrastructure  roads,1,0.42,,,0.37275441,,,0.396377205,,,,0.073943937,,,
-Infrastructure  roads,1.5,0.55,,,0.603934871,,,0.576967436,,,,0.205814323,,,
-Infrastructure  roads,2,0.65,,,0.709659091,,,0.679829545,,,,0.270555844,,,
-Infrastructure  roads,3,0.8,,,0.808409091,,,0.804204545,,,,0.221245379,,,
-Infrastructure  roads,4,0.9,,,0.887159091,,,0.893579545,,,,0.133514877,,,
-Infrastructure  roads,5,1,,,0.96875,,,0.984375,,,,0.0625,,,
-Infrastructure  roads,6,1,,,1,,,1,,,,0,,,
+Infrastructure roads,0,0,,,0,,,0,,,,0,,,
+Infrastructure roads,0.5,0.25,,,0.214436906,,,0.232218453,,,,0.069410857,,,
+Infrastructure roads,1,0.42,,,0.37275441,,,0.396377205,,,,0.073943937,,,
+Infrastructure roads,1.5,0.55,,,0.603934871,,,0.576967436,,,,0.205814323,,,
+Infrastructure roads,2,0.65,,,0.709659091,,,0.679829545,,,,0.270555844,,,
+Infrastructure roads,3,0.8,,,0.808409091,,,0.804204545,,,,0.221245379,,,
+Infrastructure roads,4,0.9,,,0.887159091,,,0.893579545,,,,0.133514877,,,
+Infrastructure roads,5,1,,,0.96875,,,0.984375,,,,0.0625,,,
+Infrastructure roads,6,1,,,1,,,1,,,,0,,,
 Agriculture,0,0,0.018575388,,0,0,,0,,,,0,0,,
 Agriculture,0.5,0.3,0.267797668,,0.135,0.242873563,,0.236417808,,,,0.188944436,0.095973845,,
 Agriculture,1,0.55,0.473677377,,0.37,0.47183908,,0.466379114,,,,0.351069794,0.17508562,,
