+++ src/physrisk/data/static/vulnerability/downtime_config.csv
