+++ src/physrisk/data/static/vulnerability/vulnerability_config.csv
