+++ src/physrisk/vulnerability_models/config_based_vuln_model_acute.py
@@ -1,0 +1,151 @@
+from typing import (
+    Iterable,
+    Sequence,
+    Tuple,
+    Union,
+)
+
+import numpy as np
+from physrisk.kernel.assets import Asset
+from physrisk.kernel.curve import ExceedanceCurve
+from physrisk.kernel.hazard_event_distrib import HazardEventDistrib
+from physrisk.kernel.hazard_model import (
+    HazardDataRequest,
+    HazardDataResponse,
+    HazardEventDataResponse,
+    HazardParameterDataResponse,
+)
+from physrisk.kernel.hazards import PluvialInundation, RiverineInundation
+from physrisk.kernel.impact_distrib import ImpactType
+from physrisk.kernel.vulnerability_distrib import VulnerabilityDistrib
+from physrisk.kernel.vulnerability_model import (
+    VulnerabilityModelAcuteBase
+)
+from physrisk.vulnerability_models.config_based_impact_curves import ImpactCurve, ImpactCurveKey, VulnerabilityConfigItem
+
+
+class ConfigBasedVulnerabilityModelAcute(VulnerabilityModelAcuteBase):
+    def __init__(
+        self,
+        *,
+        asset_class: str,  # type of the asset class
+        hazard_class: type,  # type of the hazard class
+        impact_id: str,
+        config_items: Sequence[VulnerabilityConfigItem],
+    ):
+        """A vulnerability model that requires only specification of a set of damage/disruption curves
+        which are provided by configuration. This simple model contains no uncertainty around damage/disruption.
+        The edges of the (hazard) intensity bins are determined by the granularity of the hazard data itself.
+        The impact bin edges are inferred from the intensity bin edges, by looking up the impact corresponding
+        to the hazard indicator intensity from the damage curve.
+        """
+        impact_type = (
+            ImpactType.damage if impact_id == "damage" else ImpactType.disruption
+        )
+        indicator_id = config_items[0].indicator_id
+
+        super().__init__(indicator_id, hazard_class, impact_type)
+
+        self.asset_class_type = asset_class
+        self.curves = {
+            ImpactCurveKey(item.asset_identifier): ImpactCurve.get_impact_curve(item)
+            for item in config_items
+        }
+        self.has_sop = (hazard_class == RiverineInundation) or (
+            hazard_class == PluvialInundation
+        )
+        self.asset_attributes = set(
+            key for curve in self.curves for key in curve.identifier.keys()
+        )
+
+    def get_data_requests(
+        self, asset: Asset, *, scenario: str, year: int
+    ) -> Union[HazardDataRequest, Iterable[HazardDataRequest]]:
+        future = HazardDataRequest(
+            self.hazard_type,
+            asset.longitude,
+            asset.latitude,
+            scenario=scenario,
+            year=year,
+            indicator_id=self.indicator_id,
+        )
+        if not self.has_sop:
+            return future
+        # if the acute model includes standard of protection, then we request the historical curve and (hisorical) SoP
+        # some haazrd models (e.g. JBA) will give a future-adjusted SoP, but generally we need to infer this.
+        histo = HazardDataRequest(
+            self.hazard_type,
+            asset.longitude,
+            asset.latitude,
+            scenario="historical",
+            year=year,  # will be ignored (consider making field optional?)
+            indicator_id=self.indicator_id,
+        )
+        sop = HazardDataRequest(
+            self.hazard_type,
+            asset.longitude,
+            asset.latitude,
+            scenario="historical",
+            year=year,
+            indicator_id=self._sop_id_for_hazard(self.hazard_type),
+        )
+        return [future, histo, sop]
+
+    def _sop_id_for_hazard(self, hazard_type: type):
+        if (hazard_type == RiverineInundation) or (hazard_type == PluvialInundation):
+            return "flood_sop"
+        else:
+            raise NotImplementedError(
+                f"sop not available for hazard {hazard_type.__name__}"
+            )
+
+    def get_distributions(
+        self, asset: Asset, data_responses: Iterable[HazardDataResponse]
+    ) -> Tuple[VulnerabilityDistrib, HazardEventDistrib]:
+        if self.has_sop:
+            (future, histo, sop) = data_responses
+        else:
+            (future,) = data_responses
+
+        assert isinstance(future, HazardEventDataResponse)
+        intensity_curve = ExceedanceCurve(
+            1.0 / future.return_periods, future.intensities
+        )
+
+        # sop = getattr(future, "sop", 0)  # standard of protection in years
+        if self.has_sop:
+            assert isinstance(histo, HazardEventDataResponse)
+            assert isinstance(sop, HazardParameterDataResponse)
+            if sop.parameter > 0:
+                histo_curve = ExceedanceCurve(
+                    1.0 / histo.return_periods, histo.intensities
+                )
+                # we infer protected depth from the historical curve...
+                protected_depth = histo_curve.get_value(1.0 / float(sop.parameter))
+                if protected_depth > 0:
+                    # and assume this protected depth in the future scenario (if not the same)
+                    # with a modification to underlying functions we can do this in one step, but for now:
+                    c = intensity_curve.add_value_point(protected_depth)
+                    i = np.searchsorted(c.values, protected_depth)
+                    intensity_curve = ExceedanceCurve(c.probs[i:], c.values[i:])
+
+        indicator_bin_edges, probabilities = intensity_curve.get_probability_bins(
+            include_last=True
+        )
+
+        curve = self.curves[
+            ImpactCurveKey.get(asset, self.asset_attributes, self.curves.keys())
+        ]
+        impact_bins = curve.get_impact(indicator_bin_edges)
+
+        vul = VulnerabilityDistrib(
+            type(self.hazard_type),
+            indicator_bin_edges,
+            impact_bins,
+            np.eye(len(probabilities)),
+        )
+        event = HazardEventDistrib(
+            type(self.hazard_type), indicator_bin_edges, probabilities, path=[future.path]
+        )
+
+        return vul, event
