+++ src/physrisk/vulnerability_models/downtime.py
@@ -6,7 +6,7 @@ from typing import Dict, List, Optional, Sequence, Type
 import numpy as np
 from physrisk.kernel.assets import Asset
 
-from physrisk.vulnerability_models.config_based_model import (
+from physrisk.vulnerability_models.config_based_impact_curves import (
     DowntimeConfigItem,
     ImpactCurveKey,
 )
