+++ src/physrisk/vulnerability_models/vulnerability.py
@@ -15,12 +15,12 @@ from physrisk.kernel.vulnerability_model import (
     VulnerabilityModelsFactory as PVulnerabilityModelsFactory,
 )
 import physrisk.data
-from physrisk.vulnerability_models.config_based_model import (
-    ConfigBasedVulnerabilityModel,
-    ConfigBasedVulnerabilityModelAcute,
+from physrisk.vulnerability_models.config_based_impact_curves import (
     VulnerabilityConfigItem,
     config_items_from_csv,
 )
+from physrisk.vulnerability_models.config_based_vuln_model_param import ConfigBasedVulnerabilityModel
+from physrisk.vulnerability_models.config_based_vuln_model_acute import ConfigBasedVulnerabilityModelAcute
 
 physrisk_assets = import_module("physrisk.kernel.assets")
