+++ tests/conftest.py
@@ -111,6 +111,17 @@ def vulnerability_onboarding_dir():
     return path
 
 
+@pytest.fixture(scope="session")
+def vulnerability_config_dir():
+    """Location of the vulnerability on-boarding directory..
+
+    Returns:
+        str: Directory path.
+    """
+    path = Path(__file__).parents[1] / "src" / "physrisk" / "data" / "static" / "vulnerability"
+    return path
+
+
 # def assert_expected(result: str, func_name: str, update_expected: bool):
 #     result_dict, expected_dict = get_result_expected(result, func_name, update_expected)
 #     diffs = DeepDiff(expected_dict, result_dict)
