+++ tests/vulnerability_models/config_based_vulnerability_test.py
@@ -9,7 +9,7 @@ from physrisk.api.v1.common import Assets
 from physrisk.data.inventory import EmbeddedInventory
 from physrisk.data.pregenerated_hazard_model import ZarrHazardModel
 from physrisk.data.zarr_reader import ZarrReader
-from physrisk.hazard_models.core_hazards import get_default_source_paths
+from physrisk.hazard_models.core_hazards import CoreFloodModels, CoreInventorySourcePaths, get_default_source_paths
 from physrisk.kernel.assets import (
     Asset,
     ManufacturingAsset,
@@ -26,13 +26,13 @@ from physrisk.kernel.hazards import (
     RiverineInundation,
     Wind,
 )
-from physrisk.vulnerability_models.config_based_model import (
-    ConfigBasedVulnerabilityModelAcute,
+from physrisk.vulnerability_models.config_based_impact_curves import (
     ImpactCurveKey,
     PiecewiseLinearImpactCurve,
     config_items_from_csv,
     config_items_to_csv,
 )
+from physrisk.vulnerability_models.config_based_vuln_model_acute import ConfigBasedVulnerabilityModelAcute
 from physrisk.vulnerability_models.vulnerability import (
     VulnerabilityConfigItem,
     VulnerabilityModelsFactory,
@@ -1100,13 +1100,13 @@ def config_based_vulnerability_models():
     ]
 
 
-def create_store(latitudes: Sequence[float], longitudes: Sequence[float]):
+def create_store(latitudes: Sequence[float], longitudes: Sequence[float],
+                 year: int = 2050,
+                 source_paths = get_default_source_paths()):
     # fmt: off
     flood_return_periods = [2.0, 5.0, 10.0, 25.0, 50.0, 100.0, 250.0, 500.0, 1000.0]
     wind_return_periods=[20, 30, 40, 50, 60, 70, 80, 90, 100, 200, 300, 400, 500, 600, 700, 800, 900, 1000]
 
-    # this is the best way to mock data:
-    source_paths = get_default_source_paths()
     def hazard_path(hazard_type, indicator_id, scenario, year):
         return source_paths[hazard_type](
             indicator_id=indicator_id, scenario=scenario, year=year
@@ -1122,7 +1122,7 @@ def create_store(latitudes: Sequence[float], longitudes: Sequence[float]):
         np.array([0.0012, 0.39, 0.85, 1.39, 1.75, 2.09, 2.51, 2.82, 3.12]), 
     )
     mocker.add_curves_global(
-        hazard_path(RiverineInundation, "flood_depth", "ssp585", 2050),
+        hazard_path(RiverineInundation, "flood_depth", "ssp585", year),
         longitudes,
         latitudes,
         flood_return_periods,
@@ -1132,46 +1132,46 @@ def create_store(latitudes: Sequence[float], longitudes: Sequence[float]):
         hazard_path(RiverineInundation, "flood_sop", "historical", -1),
         longitudes,
         latitudes,
-        [0.0, 0.0],
-        np.array([0.0, 0.0]), 
+        ["min", "max"],
+        np.array([100.0, 300.0]), 
     )
     mocker.add_curves_global(
-        hazard_path(CoastalInundation, "flood_depth", "ssp585", 2050),
+        hazard_path(CoastalInundation, "flood_depth", "ssp585", year),
         longitudes,
         latitudes,
         flood_return_periods,
         np.array([0., 0., 0., 0., 0., 0., 0., 0., 0.]), 
     )
     mocker.add_curves_global(
-        hazard_path(Wind, "max_speed", "ssp585", 2050),
+        hazard_path(Wind, "max_speed", "ssp585", year),
         longitudes,
         latitudes,
         wind_return_periods,
         np.array([24.44, 26.79, 28.63, 30.10, 31.03, 32.00, 33.04, 33.75, 34.23, 39.33, 43.0, 44.96, 46.19, 47.62, 48.58, 48.83, 49.79, 50.11]), 
     )
     mocker.add_curves_global(
-        hazard_path(ChronicHeat, "days_wbgt_above", "ssp585", 2050),
+        hazard_path(ChronicHeat, "days_wbgt_above", "ssp585", year),
         longitudes,
         latitudes,
         [5, 10, 15, 20, 25, 30, 35, 40, 45, 50, 55, 60],
         np.array([363.65, 350.21, 303.64, 240.48, 181.83, 128.47, 74.40, 1.40, 0.0, 0.0, 0.0, 0.0]), 
     )
     mocker.add_curves_global(
-        hazard_path(Fire, "fire_probability", "ssp585", 2050),
+        hazard_path(Fire, "fire_probability", "ssp585", year),
         longitudes,
         latitudes,
         [0],
         np.array([0.31867]), 
     )
     mocker.add_curves_global(
-        hazard_path(Drought, "months/spei3m/below/-2", "ssp585", 2050),
+        hazard_path(Drought, "months/spei3m/below/-2", "ssp585", year),
         longitudes,
         latitudes,
         [0],
         np.array([0.17918]),
     )
     mocker.add_curves_global(
-        hazard_path(Hail, "days/above/5cm", "ssp585", 2050),
+        hazard_path(Hail, "days/above/5cm", "ssp585", year),
         longitudes,
         latitudes,
         [0],
@@ -1180,7 +1180,7 @@ def create_store(latitudes: Sequence[float], longitudes: Sequence[float]):
     for temp, value in zip([25, 30, 35, 40, 45, 50, 55], 
                            [148.6, 65.31, 0.6, 0.0, 0.0, 0.0, 0.0]):
         mocker.add_curves_global(
-            hazard_path(ChronicHeat, f"days_tas/above/{temp}c", "ssp585", 2050),
+            hazard_path(ChronicHeat, f"days_tas/above/{temp}c", "ssp585", year),
             longitudes,
             latitudes,
             [0],
@@ -1308,7 +1308,7 @@ def test_config_based_vulnerability_models(working_dir):
     np.testing.assert_almost_equal(impacts["Wind"], 0.001040497)
     np.testing.assert_almost_equal(impacts["Hail"], 0.015738506)
     np.testing.assert_almost_equal(impacts["CoastalInundation"], 0.009)
-    np.testing.assert_almost_equal(impacts["RiverineInundation"], 0.101327466)
+    np.testing.assert_almost_equal(impacts["RiverineInundation"], 0.006229)
     np.testing.assert_almost_equal(impacts["Drought"], 0.0)
 
     impacts = {
@@ -1318,7 +1318,7 @@ def test_config_based_vulnerability_models(working_dir):
     np.testing.assert_almost_equal(impacts["Hail"], 0.1132266621)
     np.testing.assert_almost_equal(impacts["ChronicHeat"], 0.040246371)
     np.testing.assert_almost_equal(impacts["CoastalInundation"], 0.013)
-    np.testing.assert_almost_equal(impacts["RiverineInundation"], 0.136345865)
+    np.testing.assert_almost_equal(impacts["RiverineInundation"], 0.0080191999)
 
     impacts = {
         impact.hazard_type: impact.impact_mean
@@ -1328,7 +1328,7 @@ def test_config_based_vulnerability_models(working_dir):
     np.testing.assert_almost_equal(impacts["Drought"], 0.002986333271)
     np.testing.assert_almost_equal(impacts["Hail"], 0.003849706511)
     np.testing.assert_almost_equal(impacts["CoastalInundation"], 0.013)
-    np.testing.assert_almost_equal(impacts["RiverineInundation"], 0.136345865)
+    np.testing.assert_almost_equal(impacts["RiverineInundation"], 0.0080191999)
 
     impacts = {
         impact.hazard_type: impact.impact_mean
@@ -1338,14 +1338,14 @@ def test_config_based_vulnerability_models(working_dir):
     np.testing.assert_almost_equal(impacts["Drought"], 0.004479499906)
     np.testing.assert_almost_equal(impacts["ChronicHeat"], 0.067715752259)
     np.testing.assert_almost_equal(impacts["CoastalInundation"], 0.013)
-    np.testing.assert_almost_equal(impacts["RiverineInundation"], 0.136345865)
+    np.testing.assert_almost_equal(impacts["RiverineInundation"], 0.0080191999)
 
     impacts = {
         impact.hazard_type: impact.impact_mean
         for impact in response.asset_impacts[4].impacts
     }
     np.testing.assert_almost_equal(impacts["CoastalInundation"], 0.013)
-    np.testing.assert_almost_equal(impacts["RiverineInundation"], 0.136345865)
+    np.testing.assert_almost_equal(impacts["RiverineInundation"], 0.0080191999)
 
     impacts = {
         impact.hazard_type: impact.impact_mean
@@ -1353,7 +1353,7 @@ def test_config_based_vulnerability_models(working_dir):
     }
     np.testing.assert_almost_equal(impacts["Wind"], 0.000203201159)
     np.testing.assert_almost_equal(impacts["CoastalInundation"], 0.0)
-    np.testing.assert_almost_equal(impacts["RiverineInundation"], 0.043734252)
+    np.testing.assert_almost_equal(impacts["RiverineInundation"], 0.004)
 
     impacts = {
         impact.hazard_type: impact.impact_mean
@@ -1363,18 +1363,21 @@ def test_config_based_vulnerability_models(working_dir):
     np.testing.assert_almost_equal(impacts["Wind"], 0.000203201159)
     np.testing.assert_almost_equal(impacts["Hail"], 0.015738506029)
     np.testing.assert_almost_equal(impacts["CoastalInundation"], 0.0)
-    np.testing.assert_almost_equal(impacts["RiverineInundation"], 0.043734252)
+    np.testing.assert_almost_equal(impacts["RiverineInundation"], 0.004)
 
 
 def test_config_based_acute_model_with_sop():
-    """Test of config-based model including standard of protection information."""
-    # inunriver_rcp4p5_MIROC-ESM-CHEM_2030 flood_sop_historical_1985
+    """Test of config-based model including standard of protection information.
+    TUDelft.
+    To run as a 'live' example, just remove the mock store below
+    i.e. instantiate the ZarrHazardModel with: ZarrHazardModel(source_paths=source_paths, reader=ZarrReader())
+    """
 
     config_items = [
         VulnerabilityConfigItem(
             hazard_class="RiverineInundation",  # CoastalInundation,PluvialInundation,RiverineInundation
             asset_class="RealEstateAsset",
-            asset_identifier="type=Buildings/Residential,location=North America",
+            asset_identifier="type=Buildings/Residential,location=Europe",
             indicator_id="flood_depth",
             indicator_units="metres",
             impact_id="damage",
@@ -1398,11 +1401,12 @@ def test_config_based_acute_model_with_sop():
         )
     ]
 
-    latitudes = np.array([32.6017])
-    longitudes = np.array([-87.7811])
+    latitudes = np.array([51.7343])
+    longitudes = np.array([-1.2366])
+
     assets = [
         RealEstateAsset(
-            location="North America",
+            location="Europe",
             latitude=latitudes[0],
             longitude=longitudes[0],
             type="Buildings/Residential",
@@ -1410,32 +1414,39 @@ def test_config_based_acute_model_with_sop():
     ]
     factory = VulnerabilityModelsFactory(config=config_items)
     vulnerability_models = factory.vulnerability_models(disable_api_calls=True)
+    source_paths = CoreInventorySourcePaths(EmbeddedInventory(), CoreFloodModels.TUDelft).source_paths()
 
-    store = create_store(latitudes, longitudes)
-
-    source_paths = get_default_source_paths(EmbeddedInventory())
+    store = create_store(latitudes, longitudes, year=2085, source_paths=source_paths)
 
     request = requests.AssetImpactRequest(
         assets=Assets(items=[]),
         include_asset_level=True,
         include_calc_details=True,
-        years=[2050],
+        years=[2085], # 2050
         scenarios=["ssp585"],
     )
 
     response = requests._get_asset_impacts(
         request,
         ZarrHazardModel(source_paths=source_paths, reader=ZarrReader(store)),
+        # ZarrHazardModel(source_paths=source_paths, reader=ZarrReader()) # to convert to being a 'live' example.
         vulnerability_models=vulnerability_models,
         assets=assets,
     )
 
     impacts = {
-        impact.hazard_type: impact.impact_mean
+        impact.hazard_type: impact.impact_exceedance
         for impact in response.asset_impacts[0].impacts
     }
-    # TODO complete test
-    assert impacts is not None
+    flood_depths = [0.0012, 0.39, 0.85, 1.39, 1.75, 2.09, 2.51, 2.82, 3.12]
+    flood_return_periods = [2.0, 5.0, 10.0, 25.0, 50.0, 100.0, 250.0, 500.0, 1000.0]
+    # here the 1-in-100 year is the minimum protected level
+    # therefore we assume a defended level of 2.09 m
+    # in the pure deterministic case we can simply look up from the damage curve
+    expected_impacts = np.interp([2.09, 2.51, 2.82, 3.12], config_items[0].points_x, config_items[0].points_y)
+    expected_impacts = np.append(expected_impacts, expected_impacts[-1])
+    np.testing.assert_almost_equal(impacts["RiverineInundation"].exceed_probabilities, [0.01, 0.004, 0.002, 0.001, 0.])
+    np.testing.assert_almost_equal(impacts["RiverineInundation"].values, expected_impacts)
 
 
 @pytest.mark.skip("Only to be run when importing HAZUS damage functions.")
