+++ tests/vulnerability_models/config_generate_test.py
@@ -9,13 +9,13 @@ from physrisk.kernel.hazards import (
     RiverineInundation,
 )
 from physrisk.kernel.impact_distrib import ImpactType
-from physrisk.vulnerability_models.config_based_model import (
+from physrisk.vulnerability_models.config_based_impact_curves import (
     VulnerabilityConfigItem,
     config_items_to_csv,
 )
 
 
-def test_generate_all_config(vulnerability_onboarding_dir):
+def test_generate_all_config(vulnerability_onboarding_dir, vulnerability_config_dir):
     """Generate all vulnerability config. This is done via test because:
     - this ensures that the config remains traceable to source and
     - in general the process is not compute-intensive.
@@ -25,7 +25,7 @@ def test_generate_all_config(vulnerability_onboarding_dir):
     """
     config_items = generate_all_config(vulnerability_onboarding_dir)
     config_items_to_csv(
-        config_items, vulnerability_onboarding_dir / "candidate_config.csv"
+        config_items, vulnerability_config_dir / "candidate_vulnerability_config.csv"
     )
